Concentrate all tooltip logic inside CustomTooltip for better code hygene
